P4ADEV-2195 move methods of ConversionUtils to Utilities class
