+++ src/test/java/it/gov/pagopa/payhub/activities/util/TestUtils.java
@@ -168,7 +168,7 @@ public class TestUtils {
             @Override
             public <T> T manufacturePojo(Class<T> pojoClass, Type... genericTypeArgs) {
                 if(pojoClass.isAssignableFrom(XMLGregorianCalendar.class)) {
-                    return (T) ConversionUtils.toXMLGregorianCalendar(OffsetDateTime.now());
+                    return (T) Utilities.toXMLGregorianCalendar(OffsetDateTime.now());
                 }
                 return null;
             }
